requirements.txt
tensorflow-cpu==2.10.0
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.2
tzdata==2022.2
Pillow==9.2.0
django-storages==1.13.1
boto3==1.24.70
django-cors-headers==3.13.0
djangorestframework==3.13.1
pytz==2022.2.1